requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
bmp280==1.0.0
i2cdevice==1.0.0
pillow==11.1.0
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.6
